# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.CreateStaff:
    properties:
      active:
        type: integer
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      manager_id:
        type: integer
      phone:
        type: string
      store_id:
        type: integer
    type: object
...
  models.PatchRequest:
    properties:
      fields:
        additionalProperties: true
        type: object
      id:
        type: integer
    type: object
...
  models.StaffPrimaryKey:
    properties:
      staff_id:
        type: integer
    type: object
...
  models.TransferProductRequest:
    properties:
      from_store_id:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      to_store_id:
        type: integer
    type: object
...
  models.UpdateStaff:
    properties:
      active:
        type: integer
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      manager_id:
        type: integer
      phone:
        type: string
      staff_id:
        type: integer
      store_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /report/staff-sale:
    get:
      consumes:
      - application/json
      description: Get List StaffSale
      operationId: get_list_staff_selled
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List StaffSale
      tags:
      - TotalSelledStaff
  /report/staff-sale/{id}:
    get:
      consumes:
      - application/json
      description: total_selled_staff
      operationId: get_by_id_staff
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: total_selled_staff
      tags:
      - TotalSelledStaff
  /staff:
    get:
      consumes:
      - application/json
      description: Get List Staff
      operationId: get_list_staff
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List Staff
      tags:
      - Staff
    post:
      consumes:
      - application/json
      description: Create Staff
      operationId: create_staff
      parameters:
      - description: CreateStaffRequest
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaff'
      produces:
      - application/json
      responses:
        "201":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Staff
      tags:
      - Staff
  /staff/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Staff
      operationId: delete_staff
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: DeleteStaffRequest
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/models.StaffPrimaryKey'
      produces:
      - application/json
      responses:
        "204":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Staff
      tags:
      - Staff
    get:
      consumes:
      - application/json
      description: Get By ID Staff
      operationId: get_by_id_staff
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By ID Staff
      tags:
      - Staff
    patch:
      consumes:
      - application/json
      description: Update PATCH Staff
      operationId: update_staff
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdatePatchRequest
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/models.PatchRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update PATCH Staff
      tags:
      - Staff
    put:
      consumes:
      - application/json
      description: Update Staff
      operationId: update_staff
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateStaffRequest
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/models.UpdateStaff'
      produces:
      - application/json
      responses:
        "202":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Staff
      tags:
      - Staff
...
  /transferProduct:
    put:
      consumes:
      - application/json
      description: transfer product
      operationId: tranfer_product
      parameters:
      - description: transferProudctRequest
        in: body
        name: transferProduct
        required: true
        schema:
          $ref: '#/definitions/models.TransferProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: transfer transferProduct
      tags:
      - TransferProduct
swagger: "2.0"